requirements/base.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
